ADA-86 refactor utility, fix current academic yr function
