+++ acmw-db-app/components/FormComponents/TimeSelectionForm.jsx
@@ -1,6 +1,6 @@
 import styles from '../../styles/components/FormComponents.module.css'
 import {useState, useRef} from 'react'
-import {validateTime} from '../../pages/api/utility';
+import {validateTime} from '../../utility/utility';
 
 /*
     In order to properly record the time from the parent component, you'll need to pass in a recordTime function.
