+++ acmw-db-app/components/GHCVolunteerForm.jsx
@@ -4,7 +4,7 @@ import SelectInput from './FormComponents/SelectInput'
 import SubmitNotification from './FormComponents/SubmitNotification'
 import styles from '../styles/components/Form.module.css'
 import {useState} from 'react'
-import {validateNumVolHours} from '../pages/api/utility'
+import {validateNumVolHours} from '../utility/utility'
 
 const GHCVolunteerForm = (props) => {
